chkmadd.exp: Send MAIL FROM with empty address
Sending with the recipient address allows for a validity test
before RCPT TO, but may raise suspicion towards the mailbox owner;
we don't want that to happen, so it is disabled for now.  May be
required to be re-enabled if there is an MTA that does not accept <>
or does not say "no" on RCPT TO for not existing mailboxes.